requirements.txt
# Optional dependencies (core run-once works with Python stdlib + sqlite)
APScheduler==3.10.4
PyYAML==6.0.2
streamlit==1.37.1
